git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@14038 01821946-851b-0410-b7f4-f1477c70e57a

+++ 3d/fastmodels.cpp
@@ -580,7 +580,7 @@ if ((nExclusive < 0) || (nSubModel == nExclusive)) {
 	if (bEdges) {
 		if (!subModelP->m_bThruster) {
 			RenderModel::CModelEdge* edgeP = subModelP->m_edges.Buffer ();
-			int32_t nVertices [2] = { 0, gameData.segData.edgeVertices.Length () };
+			int32_t nVertices [2] = { 0, (int32_t) gameData.segData.edgeVertices.Length () };
 			int32_t nEdgeFilter = objP->IsWeapon () ? 0 : bHires; //0 : 2; //bHires;
 			for (i = subModelP->m_nEdges; i; i--, edgeP++) 
 				edgeP->Prepare (CFloatVector::ZERO, nVertices, nEdgeFilter);
